requirements.txt
pandas
yfinance
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
